+++ CHANGELOG.md
@@ -8,8 +8,8 @@ and this project adheres to
 
 
 ## [Unreleased]
-
-## [0.0.1] — 2019-09-09
+### Changed
+- Use `jsonista` for underlying JSON support.
 
 ## [0.0.1] — 2019-09-09
 Released without _CHANGELOG.md_.